requirements.txt
### Commented parts are already available in docker image (for local development without docker uncomment them)
h5py==3.6.0
# numpy==1.26.4
# pandas==2.1.4
Pillow==10.2.0
# torch==2.2.1 # 2.2.1+cu121
torchvision==0.17.1 # 0.17.1+cu121
ujson==5.1.0
hydra-core==1.3.2
pytorch-lightning==2.2.1
lightning==2.2.1
wandb==0.16.4
submitit==1.5.1
hydra-submitit-launcher==1.2.0
python-dotenv==1.0.1
tqdm==4.65.0
